% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-memory2.R
\name{cache_memory2}
\alias{cache_memory2}
\title{In Memory Cache with last access timestamp recorded for each object}
\usage{
cache_memory2(algo = "xxhash64")
}
\arguments{
\item{algo}{The hashing algorithm used for the cache, see
\code{\link[digest]{digest}} for available algorithms.}
}
\description{
A cache in memory, that lasts only in the current R session.
}
\details{
This function differs from \code{memoise::cache_memory} in a number of ways:

\itemize{
  \item{timestamping}  {Each get/set to the cache is timestamped so that the
       last access time for every cached object is known}
  \item{expiry}  {The addition of timestamps allows for expiring objects older
       than a certain age. See the function \code{memoisetools::expire_cache}}
  \item{faster default hasing algo}  {Use \code{xxhash64} rather than \code{sha512}}
}
}
