% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expire-cache.R
\name{expire_cache}
\alias{expire_cache}
\title{Delete objects in the cache older than the specified age}
\usage{
expire_cache(f, age_in_seconds, verbose = FALSE)
}
\arguments{
\item{f}{memoised function}

\item{age_in_seconds}{delete objects older than this age (in seconds)}

\item{verbose}{print information about the number of expired objects? default: FALSE}
}
\description{
Delete objects in the cache older than the specified age
}
