% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cache_info.R
\name{get_cache_dir_info}
\alias{get_cache_dir_info}
\title{Sum of the size of all files in a directory which have the given filename_length}
\usage{
get_cache_dir_info(path, filename_length)
}
\arguments{
\item{path}{path to cache directory}

\item{filename_length}{length of filename of files to consider}
}
\value{
list with number of objects (n) and total size of objects (bytes)
}
\description{
This function is used to get the filenames which are probably matches for the
current digest signature, as given by the filename_length
}
