% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cache_info.R
\name{get_cache_info}
\alias{get_cache_info}
\title{Get information about the memory or filesystem cache}
\usage{
get_cache_info(cache_funclist)
}
\arguments{
\item{cache_funclist}{the function list returned by a call to
memoise::cache_memory() or memoise::cache_filesystem()}
}
\value{
named list of information: \itemize{
\item{cache} - {Cache type. Either 'memory', 'filesystem' or 'gcs or aws'}
\item{env} - {For 'memory' cache, the R environment in which objects are stored}
\item{path} - {For 'filesystem' caches, the path to the cache directory}
\item{cache_name} - {For 'gcs or aws' caches, the name of the cache}
\item{algo} - {Hashing algorithm for creating keys}
\item{bytes} - {Size (in bytes) of the cache}
\item{n} - {Number of objects in the cache}
\item{has_timestamp} - {Does the cache include timestamp information?}
\item{compress} - {Value for 'compress' variable}
}
}
\description{
'gcs' and 'aws' caches not handled yet.
}
