% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cache_info.R
\name{get_memoise_info}
\alias{get_memoise_info}
\title{Get information about the cache/caches for a memoised function}
\usage{
get_memoise_info(f, verbose = TRUE)
}
\arguments{
\item{f}{memoised function}

\item{verbose}{print information about the caches}
}
\value{
(invisible) A list of lists of information for each cache used for memoisation

\itemize{
\item{cache} - {Cache type. Either 'memory' or 'filesystem'}
\item{env} - {For 'memory' cache, the R environment in which objects are stored}
\item{path} - {For 'filesystem' caches, the path to the cache directory}
\item{algo} - {Hashing algorithm for creating keys}
\item{bytes} - {Size (in bytes) of the cache}
\item{n} - {Number of objects in the cache}
\item{has_timestamp} - {Does the cache include timestamp information?}
\item{compress} - {Value for 'compress' variable}
}
}
\description{
Get information about the cache/caches for a memoised function
}
