% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoise-mixed-backend.R
\name{memoise_with_mixed_backend}
\alias{memoise_with_mixed_backend}
\title{A version of 'memoise::memoise' which does not cache results above the specified size}
\usage{
memoise_with_mixed_backend(f, ..., envir = environment(f),
  cache1 = cache_memory2(), cache2 = cache_filesystem2(),
  result_size_limit = 1048576L)
}
\arguments{
\item{f}{Function of which to create a memoised copy.}

\item{...}{optional variables specified as formulas with no RHS to use as
additional restrictions on caching. See Examples for usage.}

\item{envir}{Environment of the returned function.}

\item{cache1}{Cache function 1 - must use the same algo as cache2}

\item{cache2}{Cache function 2 - must use the same algo as cache1}

\item{result_size_limit}{maximum size (in bytes) of results stored in cache1.
Results above this size are stored in cache2. Default: 1048576 bytes (1MB)}
}
\description{
This version of memoise also tracks the last access time for each
cached result.
}
