% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoise-with-result-size-limit.R
\name{memoise_with_result_size_limit}
\alias{memoise_with_result_size_limit}
\title{A version of 'memoise::memoise' which does not cache results above the specified size}
\usage{
memoise_with_result_size_limit(f, ..., envir = environment(f),
  cache = cache_memory2(), result_size_limit = 1048576L)
}
\arguments{
\item{f}{Function of which to create a memoised copy.}

\item{...}{optional variables specified as formulas with no RHS to use as
additional restrictions on caching. See Examples for usage.}

\item{envir}{Environment of the returned function.}

\item{cache}{Cache function.#'}

\item{result_size_limit}{maximum size (in bytes) of results stored in cache.
Results above this size are never cached. Default: 1048576 bytes (1MB)}
}
\description{
This version of memoise also tracks the last access time for each
cached result.
}
